add AllFriendView
